[EMCAL-600] Move Mapper to EMCALbase
Mapper is used in simulation and reconstruction
libraries, therefore moving to EMCALbase in
order not to make EMCALSimulation dependent
on EMCALReconstruction.
