+++ Detectors/EMCAL/base/files/RCU0A.data
