+++ Detectors/EMCAL/base/files/RCU0C.data
