+++ Detectors/EMCAL/base/files/RCU1A.data
