+++ Detectors/EMCAL/base/files/RCU1C.data
